Checkpoint: pull in mpm; merge pic from Taj's version of the world

+++ src/cmd/pic/symtab.c
@@ -12,7 +12,7 @@ YYSTYPE getvar(char *s)	/* return value of variable s (usually pointer) */
 
 	p = lookup(s);
 	if (p == NULL) {
-		if (islower(s[0]))
+		if (islower((int) s[0]))
 			ERROR "no such variable as %s", s WARNING;
 		else
 			ERROR "no such place as %s", s WARNING;
